Some fixes

+++ src/prefixMessages.js
@@ -3,30 +3,38 @@
 module.exports = {
     sendMessage: async(message,content,deleteBotInterval,deleteUserInterval) => {
         try {
-       await message.channel.send(content).then(msg =>{
-        if (deleteBotInterval || deleteBotInterval > 0 || typeof deleteBotInterval != String){
-            setTimeout(() =>{
-                msg.delete()
-            },deleteBotInterval)
-        }
-        else{
-            return;
-        }
+            await message.channel.send(content).then(msg =>{
+                if (!deleteBotInterval) {
+                   return;
+                }
+                else if (deleteBotInterval || deleteBotInterval > 0 || typeof deleteBotInterval != String){
+                    setTimeout(() =>{
+                        msg.delete()
+                    },deleteBotInterval)
+                }
+                else{
+                    return;
+                }
    
-        if (deleteUserInterval || deleteUserInterval >= 0 || typeof deleteUserInterval != String){
-            setTimeout(() =>{
-                message.delete()
-            },deleteUserInterval)
-        }
-        else if (deleteUserInterval == true){
-            message.delete()
-        }
-        else{
-           return;
-        }
-     
-       })
-        }
+                if (!deleteUserInterval){
+                   return;
+                }
+                else if (deleteUserInterval || deleteUserInterval > 0 || typeof deleteUserInterval != String){
+                    setTimeout(() =>{
+                        message.delete()
+                    },deleteUserInterval)
+                }
+                else if (deleteUserInterval == true){
+       
+                    message.delete()
+                }
+                else{
+                   return;
+                }
+                
+             
+               })
+                }
         catch (e){
              console.error(e)
         }
@@ -34,7 +42,10 @@ module.exports = {
     replyMessage : async(message,content,deleteBotInterval,deleteUserInterval) => {
         try {
            await message.reply(content).then(msg =>{
-             if (deleteBotInterval || deleteBotInterval > 0 || typeof deleteBotInterval != String){
+             if (!deleteBotInterval) {
+                return;
+             }
+             else if (deleteBotInterval || deleteBotInterval > 0 || typeof deleteBotInterval != String){
                  setTimeout(() =>{
                      msg.delete()
                  },deleteBotInterval)
@@ -42,8 +53,11 @@ module.exports = {
              else{
                  return;
              }
-         
-             if (deleteUserInterval || deleteUserInterval > 0 || typeof deleteUserInterval != String){
+
+             if (!deleteUserInterval){
+                return;
+             }
+             else if (deleteUserInterval || deleteUserInterval > 0 || typeof deleteUserInterval != String){
                  setTimeout(() =>{
                      message.delete()
                  },deleteUserInterval)
